Move URL parameter check to inside useUpdatePoolInstances()
This ensures that the Pool list is updated before fetchAndSetPools() is called, allowing the imported Pools to be populated.
