+++ archetypes/Pools/index.tsx
@@ -20,7 +20,6 @@ import MintBurnModal from './MintBurnModal';
 import PoolsTable from './PoolsTable';
 import { browseReducer, BrowseState, BrowseTableRowData, DeltaEnum, RebalanceEnum } from './state';
 import * as Styles from './styles';
-import AddAltPoolFromURL from '~/archetypes/Pools/AddAltPoolFromURL';
 
 export const Browse: React.FC = () => {
     const account = useStore(selectAccount);
@@ -203,7 +202,6 @@ export const Browse: React.FC = () => {
                     sortedFilteredTokens={filteredTokens}
                 />
             )}
-            <AddAltPoolFromURL sortedFilteredTokens={filteredTokens} />
         </>
     );
 };
