+++ store/PendingCommitSlice/types.ts
@@ -4,6 +4,6 @@ export interface IPendingCommitSlice {
     commits: Record<string, Record<string, PendingCommitInfo>>;
     userPending: Record<string, Record<string, PendingCommitAmounts>>;
     addCommit: (commitInfo: PendingCommitInfo) => void;
-    addMutlipleCommits: (commits: PendingCommitInfo[]) => void;
+    addMultipleCommits: (commits: PendingCommitInfo[]) => void;
     removeCommits: (pool: string, updateIntervalId: number) => void;
 }
